vermi-backend/requirements.txt
psycopg2-binary
python-dotenv
schedule